create staff permission mixin
